% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis.R
\name{circle.anti_join}
\alias{circle.anti_join}
\title{Doughnut Join}
\usage{
circle.anti_join(outerdf, innerdf)
}
\arguments{
\item{outerdf}{outer df}

\item{innerdf}{inner df}
}
\description{
anti_join
what's in outer but not inner
}
\examples{

circle.anti_join()

}
\keyword{circle.anti_join}
