% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_functions.R
\name{compute.cdf}
\alias{compute.cdf}
\title{Compute CDF from timeseries}
\usage{
\method{compute}{cdf}(time, index)
}
\arguments{
\item{time}{Vector of timestamps, e.g. years}

\item{index}{Vector of Index values}
}
\description{
first implementation
}
\examples{
compute.cdf(time,idex)

}
\keyword{compute.cdf}
