% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/payout_functions.R
\name{deal.1.payout}
\alias{deal.1.payout}
\title{Deal 1 payout - Example}
\usage{
deal.1.payout(s, strike, exit, tick)
}
\arguments{
\item{s}{simulation vector}

\item{strike}{strike}

\item{exit}{exit}

\item{tick}{tick value}
}
\description{
Simple Strike, Exit and Tick value
}
\details{
first implementation
}
\examples{
deal.1.payout(s,strike,exit,tick)

}
\keyword{linear.payout.1}
