% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{estimate.gamma_mm}
\alias{estimate.gamma_mm}
\title{Provides Gamma method of moments estimates}
\usage{
estimate.gamma_mm(index)
}
\arguments{
\item{index}{Vector of Index values}
}
\description{
Provides Gamma method of moments estimates
}
\examples{
estimate.gamma_mm(idex)

}
\keyword{estimate.gamma_mm}
