% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{estimate.lognormal}
\alias{estimate.lognormal}
\title{Provide log-normal MLE}
\usage{
estimate.lognormal(index)
}
\arguments{
\item{index}{Vector of Index values}
}
\description{
Provide log-normal MLE
}
\examples{
estimate.lognormal(idex)

}
\keyword{estimate.lognormal}
