% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{fit.distribution}
\alias{fit.distribution}
\title{Fits selected distribution, providing a CDF graph and the vector of the parameters}
\usage{
fit.distribution(index, distribution, zero_mass = F, DOPLOT = F)
}
\arguments{
\item{index}{Vector of Index values}

\item{zero_mass}{Boolean, zero mass distribution. It defaults to FALSE}

\item{DOPLOT}{Boolean, whether to plot}
}
\description{
Fits selected distribution, providing a CDF graph and the vector of the parameters
}
\examples{
fit.distribution(index,distribution,zero_mass,DOPLOT)

}
\keyword{fit.distribution}
