% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{linear.detrending}
\alias{linear.detrending}
\title{Requires vectors describing a time serie to
provide a linearly detrended time series}
\usage{
linear.detrending(time, index, DOPLOT = F)
}
\arguments{
\item{time}{Vector of timestamps, e.g. years}

\item{index}{Vector of Index values}

\item{DOPLOT}{Boolean, whether to plot or not. It defauts to False}
}
\description{
Requires vectors describing a time serie to
provide a linearly detrended time series
}
\examples{
linear.detrending(time,idex,DOPLOT=F)

}
\keyword{compute.cdf}
