% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_functions.R
\name{payout.statistics}
\alias{payout.statistics}
\title{Payout statistics}
\usage{
payout.statistics(payout)
}
\arguments{
\item{payout}{payout vector}
}
\description{
first implementation
}
\examples{
payout.statistic(payout)

}
