% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_functions.R
\name{simulate.cdf}
\alias{simulate.cdf}
\title{Perform simulation from CDF}
\usage{
\method{simulate}{cdf}(x, pk, N = 1e+05)
}
\arguments{
\item{x}{Vector of levels}

\item{pk}{1 minus cdf}

\item{N}{number of simulations}
}
\description{
first implementation
}
\examples{
simulate.cdf(x,cdf)

}
\keyword{simulate.cdf}
