% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis.R
\name{storm.in.circle}
\alias{storm.in.circle}
\title{Storm In a Circle}
\usage{
storm.in.circle(ibtracs, insured_lat, insured_lon, circlesize,
  circleunits = "miles")
}
\arguments{
\item{insured_lat}{lat centre}

\item{insured_lon}{lon centre}

\item{circlesize}{radius of circle}

\item{circleunits}{radius units. It defaults to miles}

\item{ibracs}{shapefile with track data}
}
\description{
Storm In a Circle
}
\examples{

storm.in.circle()

}
\keyword{storm.in.circle}
